fix: empty screen refactor: add infrastructure directory
