+++ deck/ru/deck.mdx
@@ -551,14 +551,14 @@ Clean Architecture*](https://habr.com/ru/company/mobileup/blog/335382/).
 
 <Step>
 
-```ts file=../../src/ui/App.tsx title="App.tsx"
+```ts file=../../src/infrastructure/ui/App.tsx title="App.tsx"
 ```
 
 </Step>
 
 <Step>
 
-```ts file=../../src/ui/Counter.tsx title="Counter.tsx"
+```ts file=../../src/infrastructure/ui/Counter.tsx title="Counter.tsx"
 ```
 
 </Step>
