--- src/index.ts
@@ -1,3 +1,0 @@
-export * from './adapters'
-export * from './domain'
-export * from './ui'
